+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 - Add teaser to plugin and restore the previous purpose
 - Update the plugin to use the climatoology version 7.0.0
 
+- Assigning new artifact tags instead of the default version (primary/non-primary) ([#106](https://gitlab.heigit.org/climate-action/plugins/ghg-emission-from-lulc-change/-/issues/106))
+
 ## [2.3.3](https://gitlab.heigit.org/climate-action/plugins/ghg-emission-from-lulc-change/-/releases/2.3.3) - 2025-06-13
 
 ### Fixed